+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ### v4.10.8.1
 
-- 自定义变量支持正则表达式，以此衍生的自定义函数功能
+- 新增发送表情命令：`>sendEmoji(emojiID)`、`>sendRoomEmoji(roomID, emojiID)`
+- 新增自定义变量支持正则表达式，以此衍生的自定义函数功能
 > 示例：自定义发送表情的函数，自定义变量添加：
 > ```
 > >发送表情\((.+?)\) = >postData(https://api.live.bilibili.com/msg/send, bubble=0&msg=\1&color=16777215&mode=1&dm_type=1&fontsize=25&rnd=1657851774&roomid=%room_id%&csrf=%csrf%&csrf_token=%csrf%)